+++ SolaERP.Infrastructure/Entities/Order/OrderDetailLoadDto.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace SolaERP.Application.Entities.Order;
 
 public class OrderDetailLoadDto
@@ -37,9 +39,9 @@ public class OrderDetailLoadDto
     public decimal UnitPrice { get; set; }
     public decimal TotalAmount { get; set; }
     public decimal BaseRate { get; set; }
-    public int BaseMultplyDivide { get; set; }
+    public Int16 BaseMultplyDivide { get; set; }
     public decimal ReportingRate { get; set; }
-    public int ReportingMultplyDivide { get; set; }
+    public Int16 ReportingMultplyDivide { get; set; }
     public decimal BaseTotal { get; set; }
     public decimal ReportingTotal { get; set; }
     public int DiscountType { get; set; }
